+++ Documentation/Changelog.md
@@ -12,6 +12,7 @@ Release date: not released yet
 - Added `System.Threading.Lock` support to `Gapotchenko.FX.Threading.EvaluateOnce<T>` primitive
 - Added `System.Threading.Lock` support to `Gapotchenko.FX.Threading.ExecuteOnce` primitive
 - `Gapotchenko.FX.Reflection.Loader` module made as lazy as possible to avoid chicken and egg pitfalls
+- Added `Gapotchenko.FX.IOptional` interface  for the `Gapotchenko.FX.Optional<T>` type to allow an untyped value introspection
 
 ### Gapotchenko FX 2024.1